svn conversion oprava nastaveni mluvciho - doubleclick do prostoru listboxu mohl zpusobit pad pokud nebyl vybrany mluvci... (treba scroller na boku)

+++ WpfApplication2/UI/WinSetup.xaml.cs
@@ -88,7 +88,7 @@ namespace NanoTrans
                 //prehravani
 
                 UpDownSpeed.Value =  (decimal)aNastaveni.ZpomalenePrehravaniRychlost;
-                UpDownJump.Value = (int)(aNastaveni.VlnaMalySkok);
+                UpDownJump.Value = (decimal)(aNastaveni.VlnaMalySkok);
 
                 checkBoxSaveShortened.IsChecked = aNastaveni.SaveInShortFormat;
             }
